Implement possibility to specify custom prefix for page object template files, project structure refactoring, add debug configurations
