--- core/src/main/java/com/github/kotvertolet/podegen/data/PageObjectTemplate.java
@@ -1,6 +1,0 @@
-package com.github.kotvertolet.podegen.data;
-
-import java.util.List;
-
-public record PageObjectTemplate(String className, String packages, List<Element> elements) {
-}
