--- core/src/main/java/com/github/kotvertolet/podegen/parsers/Parser.java
@@ -1,9 +1,0 @@
-package com.github.kotvertolet.podegen.parsers;
-
-import com.github.kotvertolet.podegen.data.PageObjectTemplate;
-import io.github.classgraph.Resource;
-
-public interface Parser {
-
-    PageObjectTemplate parse(Resource resource);
-}
