+++ podegen-core/src/main/java/com/github/kotvertolet/podegen/core/data/PageObjectTemplate.java
@@ -1,0 +1,6 @@
+package com.github.kotvertolet.podegen.core.data;
+
+import java.util.List;
+
+public record PageObjectTemplate(String className, String packages, List<Element> elements) {
+}
