+++ podegen-core/src/main/java/com/github/kotvertolet/podegen/core/data/enums/Extension.java
@@ -1,0 +1,33 @@
+package com.github.kotvertolet.podegen.core.data.enums;
+
+import java.util.HashMap;
+import java.util.Map;
+
+public enum Extension {
+
+    JSON("json"),
+    YAML("yaml"),
+    YML("yml");
+
+    private static final Map<String, Extension> lookup = new HashMap<String, Extension>();
+
+    static {
+        for (Extension ext : Extension.values()) {
+            lookup.put(ext.getValue(), ext);
+        }
+    }
+
+    String value;
+
+    Extension(String value) {
+        this.value = value;
+    }
+
+    public static Extension get(String extension) {
+        return lookup.get(extension);
+    }
+
+    public String getValue() {
+        return value;
+    }
+}
