+++ podegen-core/src/main/java/com/github/kotvertolet/podegen/core/parsers/Parser.java
@@ -1,0 +1,9 @@
+package com.github.kotvertolet.podegen.core.parsers;
+
+import com.github.kotvertolet.podegen.core.data.PageObjectTemplate;
+import io.github.classgraph.Resource;
+
+public interface Parser {
+
+    PageObjectTemplate parse(Resource resource);
+}
