+++ podegen-core/src/main/java/com/github/kotvertolet/podegen/core/utils/StringUtils.java
@@ -1,0 +1,43 @@
+package com.github.kotvertolet.podegen.core.utils;
+
+public class StringUtils {
+
+    public static String capitalizeStr(String string) {
+        char[] chars = string.toLowerCase().toCharArray();
+        boolean found = false;
+        for (int i = 0; i < chars.length; i++) {
+            if (!found && Character.isLetter(chars[i])) {
+                chars[i] = Character.toUpperCase(chars[i]);
+                found = true;
+            } else if (Character.isWhitespace(chars[i]) || chars[i] == '.' || chars[i] == '\'') { // You can add other chars here
+                found = false;
+            }
+        }
+        return String.valueOf(chars);
+    }
+
+    public static String capitalize(String str) {
+        int strLen = str != null ? str.length() : 0;
+        if (strLen == 0) {
+            return str;
+        } else {
+            int firstCodepoint = str.codePointAt(0);
+            int newCodePoint = Character.toTitleCase(firstCodepoint);
+            if (firstCodepoint == newCodePoint) {
+                return str;
+            } else {
+                int[] newCodePoints = new int[strLen];
+                int outOffset = 0;
+                newCodePoints[outOffset++] = newCodePoint;
+
+                int codepoint;
+                for (int inOffset = Character.charCount(firstCodepoint); inOffset < strLen; inOffset += Character.charCount(codepoint)) {
+                    codepoint = str.codePointAt(inOffset);
+                    newCodePoints[outOffset++] = codepoint;
+                }
+
+                return new String(newCodePoints, 0, outOffset);
+            }
+        }
+    }
+}
