+++ pom.xml
@@ -9,8 +9,8 @@
     <packaging>pom</packaging>
     <version>1.0-SNAPSHOT</version>
     <modules>
-        <module>tests</module>
-        <module>core</module>
+        <module>podegen-tests</module>
+        <module>podegen-core</module>
     </modules>
 
     <properties>
@@ -35,12 +35,6 @@
             <artifactId>selenide</artifactId>
             <version>7.0.2</version>
             <scope>compile</scope>
-        </dependency>
-
-        <dependency>
-            <groupId>org.apache.commons</groupId>
-            <artifactId>commons-lang3</artifactId>
-            <version>3.12.0</version>
         </dependency>
     </dependencies>
 </project>
