--- tests/src/test/java/com/github/kotvertolet/podegen/Main.java
@@ -1,14 +1,0 @@
-package com.github.kotvertolet.podegen;
-
-
-import com.github.kotvertolet.podegen.annotations.PageObject;
-import com.github.kotvertolet.podegen.data.enums.Flavours;
-import com.github.kotvertolet.podegen.data.enums.Strategies;
-
-@PageObject(flavour = Flavours.Selenium, strategy = Strategies.PageObject)
-public class Main {
-
-    public static void main(String[] args) {
-
-    }
-}
